separate out the "boring" impls of `Fold` into their own module
